Move conneciton manager implementation to its own src file
