+++ src/Makefile.subdir
@@ -150,6 +150,7 @@ src_libcodes_la_SOURCES =  \
 	src/util/jobmap-impl/jobmap-identity.c\
 	src/util/codes-mapping-context.c \
   	src/util/codes-comm.c \
+	src/util/connection-manager.C \
     src/workload/codes-workload.c \
     src/workload/methods/codes-iolang-wrkld.c \
     src/workload/methods/codes-checkpoint-wrkld.c \
